// ProKart/angularapp/src/app/models/review.model.ts
import { User } from './user.model';
import { Product } from './product.model';

export interface Review {
  reviewId?: number; // Optional, as it will be generated by the backend
  reviewText: string;
  rating: number;
  date: string; // Storing date as a string for easy API compatibility
  user: User;
  product: Product;
}